move bundleItemName to utils to be shared
