+++ apps/stardew.app/src/components/cards/bundle-item-card.tsx
@@ -5,8 +5,10 @@ import { Dispatch, SetStateAction } from "react";
 import { usePlayers } from "@/contexts/players-context";
 
 import { categoryIcons, goldIcons } from "@/lib/constants";
-import { BundleItem, BundleItemWithLocation } from "@/types/bundles";
+import { BundleItemWithLocation } from "@/types/bundles";
 import { BooleanCard } from "./boolean-card";
+
+import { categoryItems } from "@/lib/utils";
 
 interface BundleItemCardProps {
 	item: BundleItemWithLocation;
@@ -27,23 +29,6 @@ interface BundleItemCardProps {
 	 * @memberof BundleItemCardProps
 	 */
 	setPromptOpen?: Dispatch<SetStateAction<boolean>>;
-}
-
-const categoryItems: Record<string, string> = {
-	"-4": "Any Fish",
-	"-5": "Any Egg",
-	"-6": "Any Milk",
-	"-777": "Wild Seeds (Any)",
-};
-
-export function bundleItemName<T extends BundleItem>(item: T): string {
-	if (item.itemID == "-1") {
-		return "Gold";
-	} else if (item.itemID in categoryItems) {
-		return categoryItems[item.itemID];
-	}
-
-	return objects[item.itemID as keyof typeof objects]?.name;
 }
 
 export const BundleItemCard = ({
